+++ BALSAMIC/workflows/QC.smk
@@ -11,7 +11,8 @@ from snakemake.exceptions import RuleException, WorkflowError
 
 from BALSAMIC.utils.exc import BalsamicError
 
-from BALSAMIC.utils.cli import (write_json, check_executable, generate_h5)
+from BALSAMIC.utils.cli import (check_executable, generate_h5)
+from BALSAMIC.utils.io import write_json
 
 from BALSAMIC.utils.models import BalsamicWorkflowConfig
